refactor: improve transaction handling and UI components; add user authorization check and streamline AddTransactionButton functionality
