+++ app/_components/add-transation-button.tsx
@@ -1,231 +1,23 @@
 "use client";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "./ui/dialog";
-import {
-  TransactionCategory,
-  TransactionPaymentMethod,
-  TransactionType,
-} from "@prisma/client";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "./ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "./ui/select";
-import {
-  TRANSACTION_CATEGORY_OPTIONS,
-  TRANSACTION_TYPE_OPTIONS,
-} from "../_constants/transations";
+
 import { ArrowDownUpIcon } from "lucide-react";
-import React, { useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
+import { useState } from "react";
 
-import { zodResolver } from "@hookform/resolvers/zod";
-
-import { addTransaction } from "../_actions/add-transation";
 import { Button } from "./ui/button";
-import { DatePicker } from "./ui/date-picker";
-import { Input } from "./ui/input";
-import { MoneyInput } from "./money-input";
+import { UpsertTransationDialog } from "./upsert-transation-dialog";
 
 export const AddTransationButton = () => {
-  const formSchema = z.object({
-    name: z.string().trim().min(1, {
-      message: "O nome e obrigatório.",
-    }),
-    type: z.nativeEnum(TransactionType, {
-      required_error: "O tipo é obrigatório",
-    }),
-    amount: z
-      .number({ message: "O valor é obrigatório" })
-      .positive("O valor deve ser positivo"),
-    category: z.nativeEnum(TransactionCategory),
-    paymentMethod: z.nativeEnum(TransactionPaymentMethod),
-    date: z.date({
-      required_error: "A data é obrigatória",
-    }),
-  });
-  const [isOpenDialog, setIsOpenDialog] = useState(false);
-
-  type FormSchema = z.infer<typeof formSchema>;
-
-  // const { handleSubmit, register } = useForm<formSchema>({});
-  const form = useForm<FormSchema>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      amount: 0,
-      type: TransactionType.EXPENSE,
-      category: TransactionCategory.OTHER,
-      paymentMethod: TransactionPaymentMethod.CASH,
-      date: new Date(),
-    },
-  });
-
-  const onSubmit = (data: FormSchema) => {
-    console.log("submit", { data });
-    try {
-      addTransaction(data);
-      setIsOpenDialog(false);
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  const [dialogIsOpen, setDialogeIsOpden] = useState(false);
 
   return (
-    <Dialog
-      open={isOpenDialog}
-      onOpenChange={(open) => {
-        setIsOpenDialog(open);
-        if (!open) {
-          form.reset();
-        }
-      }}
-    >
-      <DialogTrigger asChild>
-        <Button className="rounded-full">
-          Adicionar transação <ArrowDownUpIcon />
-        </Button>
-      </DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Adicionar transação</DialogTitle>
-          <DialogDescription>Insira as informações abaixo</DialogDescription>
-        </DialogHeader>
-
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-            <FormField
-              control={form.control}
-              name="name"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Name</FormLabel>
-                  <FormControl>
-                    <Input placeholder="Digite o nome..." {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="amount"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Amount</FormLabel>
-                  <FormControl>
-                    <MoneyInput
-                      placeholder="Digite o valor..."
-                      value={field.value}
-                      onValueChange={({ floatValue }) =>
-                        field.onChange(floatValue)
-                      }
-                      onBlur={field.onBlur}
-                      disabled={field.disabled}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="type"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Tipo</FormLabel>
-                  <Select
-                    onValueChange={field.onChange}
-                    defaultValue={field.value}
-                  >
-                    <FormControl>
-                      <SelectTrigger>
-                        <SelectValue placeholder="Selecione o tipo " />
-                      </SelectTrigger>
-                    </FormControl>
-                    <SelectContent>
-                      {TRANSACTION_TYPE_OPTIONS.map((ptions) => (
-                        <SelectItem key={ptions.value} value={ptions.value}>
-                          {ptions.label}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="category"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Categoria</FormLabel>
-                  <Select
-                    onValueChange={field.onChange}
-                    defaultValue={field.value}
-                  >
-                    <FormControl>
-                      <SelectTrigger>
-                        <SelectValue placeholder="Selecione a categoria..." />
-                      </SelectTrigger>
-                    </FormControl>
-                    <SelectContent>
-                      {TRANSACTION_CATEGORY_OPTIONS.map((ptions) => (
-                        <SelectItem key={ptions.value} value={ptions.value}>
-                          {ptions.label}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="date"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Data</FormLabel>
-                  <DatePicker value={field.value} onChange={field.onChange} />
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <DialogFooter>
-              <DialogClose asChild>
-                <Button type="button" variant="outline">
-                  Cancelar
-                </Button>
-              </DialogClose>
-              <Button disabled={form.formState.isSubmitting} type="submit">
-                Adicionar
-              </Button>
-            </DialogFooter>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
+    <>
+      <Button className="rounded-full" onClick={() => setDialogeIsOpden(true)}>
+        Adicionar transação <ArrowDownUpIcon />
+      </Button>
+      <UpsertTransationDialog
+        isOpen={dialogIsOpen}
+        setIsOpen={setDialogeIsOpden}
+      />
+    </>
   );
 };
